+++ middleware/error-handling.ts
@@ -1,0 +1,14 @@
+import { Request, Response, NextFunction } from 'express';
+
+import { ApiError } from '../error-handler/error-handler'
+
+export const errorHandlingMiddleware = function (error: Error, request: Request, response: Response, next: NextFunction): Response {
+  if (error instanceof ApiError) {
+    return response.status(error.status).json({
+      message: error.message
+    });
+  }
+  return response.status(500).json({
+    message: "Unexpected Error"
+  });
+};
