+++ package.json
@@ -5,8 +5,8 @@
   "author": "the best team ever",
   "license": "MIT",
   "scripts": {
-    "start": "cross-env NODE_ENV=production node server.js",
-    "server": "cross-env NODE_ENV=development nodemon server.js",
+    "start": "cross-env NODE_ENV=production node server.ts",
+    "server": "cross-env NODE_ENV=development nodemon server.ts",
     "client": "yarn --cwd client run start",
     "client:install": "yarn --cwd client add react-scripts --save",
     "client:build": "yarn --cwd client run build",
@@ -14,8 +14,14 @@
     "heroku-postbuild": "yarn client:install && yarn client:build"
   },
   "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
+    "@types/express": "^4.17.16",
+    "@types/jsonwebtoken": "^9.0.1",
+    "@types/node": "^18.11.18",
     "concurrently": "^7.6.0",
-    "nodemon": "^2.0.20"
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
