--- routes/index.js
@@ -1,8 +1,0 @@
-const Router = require("express");
-
-const router = new Router();
-const userRouter = require("./user-router");
-
-router.use("/user", userRouter);
-
-module.exports = router;
