+++ routes/index.ts
@@ -1,0 +1,6 @@
+import express from "express";
+
+import { userRouter } from './user-router';
+
+export const router = express.Router();
+router.use("/user", userRouter);
